pew pew

+++ First Person Shooter/Assets/Scripts/Interfaces/IDamageable.cs
@@ -3,5 +3,5 @@ using UnityEngine;
 public interface IDamageable
 {
     // Use this on every script that handles damage for any object
-    void Damage(float damageAmount, Vector3 hitPos);
+    void Damage(int damageAmount, Vector3 hitPos);
 }
